+++ src/base/core/src/lang/repr/mod.rs
@@ -2,6 +2,7 @@
 //
 #![doc = crate::_TAG_WIP!()]
 //! Representation languages.
+#![doc = crate::_doc!(flat:"lang")]
 //!
 //! Declarative languages for describing data, documents, and structure.
 //