+++ src/base/core/src/lang/sem/mod.rs
@@ -2,6 +2,7 @@
 //
 #![doc = crate::_TAG_WIP!()]
 //! Semantic relations, independent of form and execution.
+#![doc = crate::_doc!(flat:"lang")]
 //!
 //! > Meaning is relational.
 //