+++ src/lang/gram/mod.rs
@@ -12,6 +12,5 @@
 //
 
 crate::structural_mods! { // _mods
-    _mods {
-    }
+    _mods {}
 }