+++ src/lang/prog/mod.rs
@@ -1,8 +1,9 @@
 // devela::lang::prog
 //
 #![doc = crate::_DOC_LANG_PROG!()]
-#![doc = crate::_doc!(modules: crate::lang; prog: ffi)] // dsl
+#![doc = crate::_doc!(modules: crate::lang; prog: ffi)] // dsl, ir
 #![doc = crate::_doc!(flat:"lang")]
+#![doc = crate::_doc!(newline)]
 //!
 #![doc = crate::_doc!(extends: ffi)]
 //!