+++ src/phys/astro/mod.rs
@@ -1,0 +1,7 @@
+// devela::phys::astro
+//
+#![doc = crate::_TAG_WIP!()]
+//! Astronomy, celestial mechanics.
+//
+
+crate::structural_mods! {}