+++ src/phys/mod.rs
@@ -11,7 +11,7 @@
 #![cfg_attr(feature = "safe_phys", forbid(unsafe_code))]
 // docs
 crate::CONST! { pub(crate) _DOC_PHYS_MODULES =
-    crate::_doc!(modules: crate; phys: bio, chem, elec, mech, time, unit, wave);
+    crate::_doc!(modules: crate; phys: astro, bio, chem, elec, mech, time, unit, wave);
 }
 
 pub mod bio;