+++ src/run/frame/mod.rs
@@ -1,5 +1,6 @@
 // devela::run::frame
 //
+#![doc = crate::_TAG_WIP!()]
 #![doc = crate::_DOC_RUN_FRAME!()]
 //
 
@@ -8,7 +9,7 @@
 crate::structural_mods! { // _mods
     _mods {
         // pub use super::{
-        //     pacer::*, // WIP
+        //     pacer::*,
         // };
     }
 }